+++ NEWS.md
@@ -1,9 +1,25 @@
 # ClinAD 0.1.0-1
 
-## New
+## Data
 
 * New data: `ae`, `dm`, `eg`, `lb`, `mh`, `pe`, `vs`.
+
+## Function
+
+* New class `Domain`.
+* New class `DM`.
+* New generic and method `DM`:
+    + `DM,data.frame,missing-method`.
+    + `DM,data.frame,data.frame-method`.
+
+## Tests
+
+* New: `lintr` test for code quality.
+* New: [`accessor`] select a row.
+* New: [`accessor`] select a col.
+
+## Package
+
 * New citation management.
 * New pkgdown for automated webpage.
-* New `lintr` test for code quality.
-
+* New Vignette: `dm`.